Fix various things
Add .bat & working_dirs to .gitignore
Split SpriteSelector to own class
Move Adjust to own file
Move Bottom to own file
Move Custom to own file
Move Randomizer pages to own files

Likely with memory leaks, yay.
