+++ classes/__init__.py
@@ -1,0 +1,1 @@
+# do nothing, just exist to make "classes" package
