+++ gui/custom/overview.py
@@ -1,0 +1,593 @@
+from tkinter import ttk, filedialog, messagebox, IntVar, StringVar, Button, Checkbutton, Entry, Frame, Label, OptionMenu, Spinbox, E, W, LEFT, RIGHT, X
+
+def custom_page(top,parent):
+    self = ttk.Frame(parent)
+
+    def validation(P):
+        if str.isdigit(P) or P == "":
+            return True
+        else:
+            return False
+    vcmd=(self.register(validation), '%P')
+
+    itemList1 = Frame(self)
+    itemList2 = Frame(self)
+    itemList3 = Frame(self)
+    itemList4 = Frame(self)
+    itemList5 = Frame(self)
+
+    bowFrame = Frame(itemList1)
+    bowLabel = Label(bowFrame, text='Bow')
+    self.bowVar = StringVar(value='0')
+    bowEntry = Entry(bowFrame, textvariable=self.bowVar, width=3, validate='all', vcmd=vcmd)
+    bowFrame.pack()
+    bowLabel.pack(anchor=W, side=LEFT, padx=(0,53))
+    bowEntry.pack(anchor=E)
+
+    progbowFrame = Frame(itemList1)
+    progbowLabel = Label(progbowFrame, text='Prog.Bow')
+    self.progbowVar = StringVar(value='2')
+    progbowEntry = Entry(progbowFrame, textvariable=self.progbowVar, width=3, validate='all', vcmd=vcmd)
+    progbowFrame.pack()
+    progbowLabel.pack(anchor=W, side=LEFT, padx=(0,25))
+    progbowEntry.pack(anchor=E)
+
+    boomerangFrame = Frame(itemList1)
+    boomerangLabel = Label(boomerangFrame, text='Boomerang')
+    self.boomerangVar = StringVar(value='1')
+    boomerangEntry = Entry(boomerangFrame, textvariable=self.boomerangVar, width=3, validate='all', vcmd=vcmd)
+    boomerangFrame.pack()
+    boomerangLabel.pack(anchor=W, side=LEFT, padx=(0,14))
+    boomerangEntry.pack(anchor=E)
+
+    magicboomerangFrame = Frame(itemList1)
+    magicboomerangLabel = Label(magicboomerangFrame, text='M.Boomerang')
+    self.magicboomerangVar = StringVar(value='1')
+    magicboomerangEntry = Entry(magicboomerangFrame, textvariable=self.magicboomerangVar, width=3, validate='all', vcmd=vcmd)
+    magicboomerangFrame.pack()
+    magicboomerangLabel.pack(anchor=W, side=LEFT)
+    magicboomerangEntry.pack(anchor=E)
+
+    hookshotFrame = Frame(itemList1)
+    hookshotLabel = Label(hookshotFrame, text='Hookshot')
+    self.hookshotVar = StringVar(value='1')
+    hookshotEntry = Entry(hookshotFrame, textvariable=self.hookshotVar, width=3, validate='all', vcmd=vcmd)
+    hookshotFrame.pack()
+    hookshotLabel.pack(anchor=W, side=LEFT, padx=(0,24))
+    hookshotEntry.pack(anchor=E)
+
+    mushroomFrame = Frame(itemList1)
+    mushroomLabel = Label(mushroomFrame, text='Mushroom')
+    self.mushroomVar = StringVar(value='1')
+    mushroomEntry = Entry(mushroomFrame, textvariable=self.mushroomVar, width=3, validate='all', vcmd=vcmd)
+    mushroomFrame.pack()
+    mushroomLabel.pack(anchor=W, side=LEFT, padx=(0,17))
+    mushroomEntry.pack(anchor=E)
+
+    magicpowderFrame = Frame(itemList1)
+    magicpowderLabel = Label(magicpowderFrame, text='Magic Powder')
+    self.magicpowderVar = StringVar(value='1')
+    magicpowderEntry = Entry(magicpowderFrame, textvariable=self.magicpowderVar, width=3, validate='all', vcmd=vcmd)
+    magicpowderFrame.pack()
+    magicpowderLabel.pack(anchor=W, side=LEFT)
+    magicpowderEntry.pack(anchor=E)
+
+    firerodFrame = Frame(itemList1)
+    firerodLabel = Label(firerodFrame, text='Fire Rod')
+    self.firerodVar = StringVar(value='1')
+    firerodEntry = Entry(firerodFrame, textvariable=self.firerodVar, width=3, validate='all', vcmd=vcmd)
+    firerodFrame.pack()
+    firerodLabel.pack(anchor=W, side=LEFT, padx=(0,33))
+    firerodEntry.pack(anchor=E)
+
+    icerodFrame = Frame(itemList1)
+    icerodLabel = Label(icerodFrame, text='Ice Rod')
+    self.icerodVar = StringVar(value='1')
+    icerodEntry = Entry(icerodFrame, textvariable=self.icerodVar, width=3, validate='all', vcmd=vcmd)
+    icerodFrame.pack()
+    icerodLabel.pack(anchor=W, side=LEFT, padx=(0,37))
+    icerodEntry.pack(anchor=E)
+
+    bombosFrame = Frame(itemList1)
+    bombosLabel = Label(bombosFrame, text='Bombos')
+    self.bombosVar = StringVar(value='1')
+    bombosEntry = Entry(bombosFrame, textvariable=self.bombosVar, width=3, validate='all', vcmd=vcmd)
+    bombosFrame.pack()
+    bombosLabel.pack(anchor=W, side=LEFT, padx=(0,32))
+    bombosEntry.pack(anchor=E)
+
+    etherFrame = Frame(itemList1)
+    etherLabel = Label(etherFrame, text='Ether')
+    self.etherVar = StringVar(value='1')
+    etherEntry = Entry(etherFrame, textvariable=self.etherVar, width=3, validate='all', vcmd=vcmd)
+    etherFrame.pack()
+    etherLabel.pack(anchor=W, side=LEFT, padx=(0,49))
+    etherEntry.pack(anchor=E)
+
+    quakeFrame = Frame(itemList1)
+    quakeLabel = Label(quakeFrame, text='Quake')
+    self.quakeVar = StringVar(value='1')
+    quakeEntry = Entry(quakeFrame, textvariable=self.quakeVar, width=3, validate='all', vcmd=vcmd)
+    quakeFrame.pack()
+    quakeLabel.pack(anchor=W, side=LEFT, padx=(0,42))
+    quakeEntry.pack(anchor=E)
+
+    lampFrame = Frame(itemList1)
+    lampLabel = Label(lampFrame, text='Lamp')
+    self.lampVar = StringVar(value='1')
+    lampEntry = Entry(lampFrame, textvariable=self.lampVar, width=3, validate='all', vcmd=vcmd)
+    lampFrame.pack()
+    lampLabel.pack(anchor=W, side=LEFT, padx=(0,46))
+    lampEntry.pack(anchor=E)
+
+    hammerFrame = Frame(itemList1)
+    hammerLabel = Label(hammerFrame, text='Hammer')
+    self.hammerVar = StringVar(value='1')
+    hammerEntry = Entry(hammerFrame, textvariable=self.hammerVar, width=3, validate='all', vcmd=vcmd)
+    hammerFrame.pack()
+    hammerLabel.pack(anchor=W, side=LEFT, padx=(0,29))
+    hammerEntry.pack(anchor=E)
+
+    shovelFrame = Frame(itemList1)
+    shovelLabel = Label(shovelFrame, text='Shovel')
+    self.shovelVar = StringVar(value='1')
+    shovelEntry = Entry(shovelFrame, textvariable=self.shovelVar, width=3, validate='all', vcmd=vcmd)
+    shovelFrame.pack()
+    shovelLabel.pack(anchor=W, side=LEFT, padx=(0,41))
+    shovelEntry.pack(anchor=E)
+
+    fluteFrame = Frame(itemList1)
+    fluteLabel = Label(fluteFrame, text='Flute')
+    self.fluteVar = StringVar(value='1')
+    fluteEntry = Entry(fluteFrame, textvariable=self.fluteVar, width=3, validate='all', vcmd=vcmd)
+    fluteFrame.pack()
+    fluteLabel.pack(anchor=W, side=LEFT, padx=(0,50))
+    fluteEntry.pack(anchor=E)
+
+    bugnetFrame = Frame(itemList2)
+    bugnetLabel = Label(bugnetFrame, text='Bug Net')
+    self.bugnetVar = StringVar(value='1')
+    bugnetEntry = Entry(bugnetFrame, textvariable=self.bugnetVar, width=3, validate='all', vcmd=vcmd)
+    bugnetFrame.pack()
+    bugnetLabel.pack(anchor=W, side=LEFT, padx=(0,41))
+    bugnetEntry.pack(anchor=E)
+
+    bookFrame = Frame(itemList2)
+    bookLabel = Label(bookFrame, text='Book')
+    self.bookVar = StringVar(value='1')
+    bookEntry = Entry(bookFrame, textvariable=self.bookVar, width=3, validate='all', vcmd=vcmd)
+    bookFrame.pack()
+    bookLabel.pack(anchor=W, side=LEFT, padx=(0,57))
+    bookEntry.pack(anchor=E)
+
+    bottleFrame = Frame(itemList2)
+    bottleLabel = Label(bottleFrame, text='Bottle')
+    self.bottleVar = StringVar(value='4')
+    bottleEntry = Entry(bottleFrame, textvariable=self.bottleVar, width=3, validate='all', vcmd=vcmd)
+    bottleFrame.pack()
+    bottleLabel.pack(anchor=W, side=LEFT, padx=(0,53))
+    bottleEntry.pack(anchor=E)
+
+    somariaFrame = Frame(itemList2)
+    somariaLabel = Label(somariaFrame, text='C.Somaria')
+    self.somariaVar = StringVar(value='1')
+    somariaEntry = Entry(somariaFrame, textvariable=self.somariaVar, width=3, validate='all', vcmd=vcmd)
+    somariaFrame.pack()
+    somariaLabel.pack(anchor=W, side=LEFT, padx=(0,30))
+    somariaEntry.pack(anchor=E)
+
+    byrnaFrame = Frame(itemList2)
+    byrnaLabel = Label(byrnaFrame, text='C.Byrna')
+    self.byrnaVar = StringVar(value='1')
+    byrnaEntry = Entry(byrnaFrame, textvariable=self.byrnaVar, width=3, validate='all', vcmd=vcmd)
+    byrnaFrame.pack()
+    byrnaLabel.pack(anchor=W, side=LEFT, padx=(0,43))
+    byrnaEntry.pack(anchor=E)
+
+    capeFrame = Frame(itemList2)
+    capeLabel = Label(capeFrame, text='Magic Cape')
+    self.capeVar = StringVar(value='1')
+    capeEntry = Entry(capeFrame, textvariable=self.capeVar, width=3, validate='all', vcmd=vcmd)
+    capeFrame.pack()
+    capeLabel.pack(anchor=W, side=LEFT, padx=(0,21))
+    capeEntry.pack(anchor=E)
+
+    mirrorFrame = Frame(itemList2)
+    mirrorLabel = Label(mirrorFrame, text='Magic Mirror')
+    self.mirrorVar = StringVar(value='1')
+    mirrorEntry = Entry(mirrorFrame, textvariable=self.mirrorVar, width=3, validate='all', vcmd=vcmd)
+    mirrorFrame.pack()
+    mirrorLabel.pack(anchor=W, side=LEFT, padx=(0,15))
+    mirrorEntry.pack(anchor=E)
+
+    bootsFrame = Frame(itemList2)
+    bootsLabel = Label(bootsFrame, text='Pegasus Boots')
+    self.bootsVar = StringVar(value='1')
+    bootsEntry = Entry(bootsFrame, textvariable=self.bootsVar, width=3, validate='all', vcmd=vcmd)
+    bootsFrame.pack()
+    bootsLabel.pack(anchor=W, side=LEFT, padx=(0,8))
+    bootsEntry.pack(anchor=E)
+
+    powergloveFrame = Frame(itemList2)
+    powergloveLabel = Label(powergloveFrame, text='Power Glove')
+    self.powergloveVar = StringVar(value='0')
+    powergloveEntry = Entry(powergloveFrame, textvariable=self.powergloveVar, width=3, validate='all', vcmd=vcmd)
+    powergloveFrame.pack()
+    powergloveLabel.pack(anchor=W, side=LEFT, padx=(0,18))
+    powergloveEntry.pack(anchor=E)
+
+    titansmittFrame = Frame(itemList2)
+    titansmittLabel = Label(titansmittFrame, text='Titan\'s Mitt')
+    self.titansmittVar = StringVar(value='0')
+    titansmittEntry = Entry(titansmittFrame, textvariable=self.titansmittVar, width=3, validate='all', vcmd=vcmd)
+    titansmittFrame.pack()
+    titansmittLabel.pack(anchor=W, side=LEFT, padx=(0,24))
+    titansmittEntry.pack(anchor=E)
+
+    proggloveFrame = Frame(itemList2)
+    proggloveLabel = Label(proggloveFrame, text='Prog.Glove')
+    self.proggloveVar = StringVar(value='2')
+    proggloveEntry = Entry(proggloveFrame, textvariable=self.proggloveVar, width=3, validate='all', vcmd=vcmd)
+    proggloveFrame.pack()
+    proggloveLabel.pack(anchor=W, side=LEFT, padx=(0,26))
+    proggloveEntry.pack(anchor=E)
+
+    flippersFrame = Frame(itemList2)
+    flippersLabel = Label(flippersFrame, text='Flippers')
+    self.flippersVar = StringVar(value='1')
+    flippersEntry = Entry(flippersFrame, textvariable=self.flippersVar, width=3, validate='all', vcmd=vcmd)
+    flippersFrame.pack()
+    flippersLabel.pack(anchor=W, side=LEFT, padx=(0,43))
+    flippersEntry.pack(anchor=E)
+
+    pearlFrame = Frame(itemList2)
+    pearlLabel = Label(pearlFrame, text='Moon Pearl')
+    self.pearlVar = StringVar(value='1')
+    pearlEntry = Entry(pearlFrame, textvariable=self.pearlVar, width=3, validate='all', vcmd=vcmd)
+    pearlFrame.pack()
+    pearlLabel.pack(anchor=W, side=LEFT, padx=(0,23))
+    pearlEntry.pack(anchor=E)
+
+    heartpieceFrame = Frame(itemList2)
+    heartpieceLabel = Label(heartpieceFrame, text='Piece of Heart')
+    self.heartpieceVar = StringVar(value='24')
+    heartpieceEntry = Entry(heartpieceFrame, textvariable=self.heartpieceVar, width=3, validate='all', vcmd=vcmd)
+    heartpieceFrame.pack()
+    heartpieceLabel.pack(anchor=W, side=LEFT, padx=(0,10))
+    heartpieceEntry.pack(anchor=E)
+
+    fullheartFrame = Frame(itemList2)
+    fullheartLabel = Label(fullheartFrame, text='Heart Container')
+    self.fullheartVar = StringVar(value='10')
+    fullheartEntry = Entry(fullheartFrame, textvariable=self.fullheartVar, width=3, validate='all', vcmd=vcmd)
+    fullheartFrame.pack()
+    fullheartLabel.pack(anchor=W, side=LEFT)
+    fullheartEntry.pack(anchor=E)
+
+    sancheartFrame = Frame(itemList2)
+    sancheartLabel = Label(sancheartFrame, text='Sanctuary Heart')
+    self.sancheartVar = StringVar(value='1')
+    sancheartEntry = Entry(sancheartFrame, textvariable=self.sancheartVar, width=3, validate='all', vcmd=vcmd)
+    sancheartFrame.pack()
+    sancheartLabel.pack(anchor=W, side=LEFT)
+    sancheartEntry.pack(anchor=E)
+
+    sword1Frame = Frame(itemList3)
+    sword1Label = Label(sword1Frame, text='Sword 1')
+    self.sword1Var = StringVar(value='0')
+    sword1Entry = Entry(sword1Frame, textvariable=self.sword1Var, width=3, validate='all', vcmd=vcmd)
+    sword1Frame.pack()
+    sword1Label.pack(anchor=W, side=LEFT, padx=(0,34))
+    sword1Entry.pack(anchor=E)
+
+    sword2Frame = Frame(itemList3)
+    sword2Label = Label(sword2Frame, text='Sword 2')
+    self.sword2Var = StringVar(value='0')
+    sword2Entry = Entry(sword2Frame, textvariable=self.sword2Var, width=3, validate='all', vcmd=vcmd)
+    sword2Frame.pack()
+    sword2Label.pack(anchor=W, side=LEFT, padx=(0,34))
+    sword2Entry.pack(anchor=E)
+
+    sword3Frame = Frame(itemList3)
+    sword3Label = Label(sword3Frame, text='Sword 3')
+    self.sword3Var = StringVar(value='0')
+    sword3Entry = Entry(sword3Frame, textvariable=self.sword3Var, width=3, validate='all', vcmd=vcmd)
+    sword3Frame.pack()
+    sword3Label.pack(anchor=W, side=LEFT, padx=(0,34))
+    sword3Entry.pack(anchor=E)
+
+    sword4Frame = Frame(itemList3)
+    sword4Label = Label(sword4Frame, text='Sword 4')
+    self.sword4Var = StringVar(value='0')
+    sword4Entry = Entry(sword4Frame, textvariable=self.sword4Var, width=3, validate='all', vcmd=vcmd)
+    sword4Frame.pack()
+    sword4Label.pack(anchor=W, side=LEFT, padx=(0,34))
+    sword4Entry.pack(anchor=E)
+
+    progswordFrame = Frame(itemList3)
+    progswordLabel = Label(progswordFrame, text='Prog.Sword')
+    self.progswordVar = StringVar(value='4')
+    progswordEntry = Entry(progswordFrame, textvariable=self.progswordVar, width=3, validate='all', vcmd=vcmd)
+    progswordFrame.pack()
+    progswordLabel.pack(anchor=W, side=LEFT, padx=(0,15))
+    progswordEntry.pack(anchor=E)
+
+    shield1Frame = Frame(itemList3)
+    shield1Label = Label(shield1Frame, text='Shield 1')
+    self.shield1Var = StringVar(value='0')
+    shield1Entry = Entry(shield1Frame, textvariable=self.shield1Var, width=3, validate='all', vcmd=vcmd)
+    shield1Frame.pack()
+    shield1Label.pack(anchor=W, side=LEFT, padx=(0,35))
+    shield1Entry.pack(anchor=E)
+
+    shield2Frame = Frame(itemList3)
+    shield2Label = Label(shield2Frame, text='Shield 2')
+    self.shield2Var = StringVar(value='0')
+    shield2Entry = Entry(shield2Frame, textvariable=self.shield2Var, width=3, validate='all', vcmd=vcmd)
+    shield2Frame.pack()
+    shield2Label.pack(anchor=W, side=LEFT, padx=(0,35))
+    shield2Entry.pack(anchor=E)
+
+    shield3Frame = Frame(itemList3)
+    shield3Label = Label(shield3Frame, text='Shield 3')
+    self.shield3Var = StringVar(value='0')
+    shield3Entry = Entry(shield3Frame, textvariable=self.shield3Var, width=3, validate='all', vcmd=vcmd)
+    shield3Frame.pack()
+    shield3Label.pack(anchor=W, side=LEFT, padx=(0,35))
+    shield3Entry.pack(anchor=E)
+
+    progshieldFrame = Frame(itemList3)
+    progshieldLabel = Label(progshieldFrame, text='Prog.Shield')
+    self.progshieldVar = StringVar(value='3')
+    progshieldEntry = Entry(progshieldFrame, textvariable=self.progshieldVar, width=3, validate='all', vcmd=vcmd)
+    progshieldFrame.pack()
+    progshieldLabel.pack(anchor=W, side=LEFT, padx=(0,16))
+    progshieldEntry.pack(anchor=E)
+
+    bluemailFrame = Frame(itemList3)
+    bluemailLabel = Label(bluemailFrame, text='Blue Mail')
+    self.bluemailVar = StringVar(value='0')
+    bluemailEntry = Entry(bluemailFrame, textvariable=self.bluemailVar, width=3, validate='all', vcmd=vcmd)
+    bluemailFrame.pack()
+    bluemailLabel.pack(anchor=W, side=LEFT, padx=(0,27))
+    bluemailEntry.pack(anchor=E)
+
+    redmailFrame = Frame(itemList3)
+    redmailLabel = Label(redmailFrame, text='Red Mail')
+    self.redmailVar = StringVar(value='0')
+    redmailEntry = Entry(redmailFrame, textvariable=self.redmailVar, width=3, validate='all', vcmd=vcmd)
+    redmailFrame.pack()
+    redmailLabel.pack(anchor=W, side=LEFT, padx=(0,30))
+    redmailEntry.pack(anchor=E)
+
+    progmailFrame = Frame(itemList3)
+    progmailLabel = Label(progmailFrame, text='Prog.Mail')
+    self.progmailVar = StringVar(value='2')
+    progmailEntry = Entry(progmailFrame, textvariable=self.progmailVar, width=3, validate='all', vcmd=vcmd)
+    progmailFrame.pack()
+    progmailLabel.pack(anchor=W, side=LEFT, padx=(0,25))
+    progmailEntry.pack(anchor=E)
+
+    halfmagicFrame = Frame(itemList3)
+    halfmagicLabel = Label(halfmagicFrame, text='Half Magic')
+    self.halfmagicVar = StringVar(value='1')
+    halfmagicEntry = Entry(halfmagicFrame, textvariable=self.halfmagicVar, width=3, validate='all', vcmd=vcmd)
+    halfmagicFrame.pack()
+    halfmagicLabel.pack(anchor=W, side=LEFT, padx=(0,18))
+    halfmagicEntry.pack(anchor=E)
+
+    quartermagicFrame = Frame(itemList3)
+    quartermagicLabel = Label(quartermagicFrame, text='Quarter Magic')
+    self.quartermagicVar = StringVar(value='0')
+    quartermagicEntry = Entry(quartermagicFrame, textvariable=self.quartermagicVar, width=3, validate='all', vcmd=vcmd)
+    quartermagicFrame.pack()
+    quartermagicLabel.pack(anchor=W, side=LEFT)
+    quartermagicEntry.pack(anchor=E)
+
+    bcap5Frame = Frame(itemList3)
+    bcap5Label = Label(bcap5Frame, text='Bomb C.+5')
+    self.bcap5Var = StringVar(value='0')
+    bcap5Entry = Entry(bcap5Frame, textvariable=self.bcap5Var, width=3, validate='all', vcmd=vcmd)
+    bcap5Frame.pack()
+    bcap5Label.pack(anchor=W, side=LEFT, padx=(0,16))
+    bcap5Entry.pack(anchor=E)
+
+    bcap10Frame = Frame(itemList3)
+    bcap10Label = Label(bcap10Frame, text='Bomb C.+10')
+    self.bcap10Var = StringVar(value='0')
+    bcap10Entry = Entry(bcap10Frame, textvariable=self.bcap10Var, width=3, validate='all', vcmd=vcmd)
+    bcap10Frame.pack()
+    bcap10Label.pack(anchor=W, side=LEFT, padx=(0,10))
+    bcap10Entry.pack(anchor=E)
+
+    acap5Frame = Frame(itemList4)
+    acap5Label = Label(acap5Frame, text='Arrow C.+5')
+    self.acap5Var = StringVar(value='0')
+    acap5Entry = Entry(acap5Frame, textvariable=self.acap5Var, width=3, validate='all', vcmd=vcmd)
+    acap5Frame.pack()
+    acap5Label.pack(anchor=W, side=LEFT, padx=(0,7))
+    acap5Entry.pack(anchor=E)
+
+    acap10Frame = Frame(itemList4)
+    acap10Label = Label(acap10Frame, text='Arrow C.+10')
+    self.acap10Var = StringVar(value='0')
+    acap10Entry = Entry(acap10Frame, textvariable=self.acap10Var, width=3, validate='all', vcmd=vcmd)
+    acap10Frame.pack()
+    acap10Label.pack(anchor=W, side=LEFT, padx=(0,1))
+    acap10Entry.pack(anchor=E)
+
+    arrow1Frame = Frame(itemList4)
+    arrow1Label = Label(arrow1Frame, text='Arrow (1)')
+    self.arrow1Var = StringVar(value='1')
+    arrow1Entry = Entry(arrow1Frame, textvariable=self.arrow1Var, width=3, validate='all', vcmd=vcmd)
+    arrow1Frame.pack()
+    arrow1Label.pack(anchor=W, side=LEFT, padx=(0,18))
+    arrow1Entry.pack(anchor=E)
+
+    arrow10Frame = Frame(itemList4)
+    arrow10Label = Label(arrow10Frame, text='Arrows (10)')
+    self.arrow10Var = StringVar(value='12')
+    arrow10Entry = Entry(arrow10Frame, textvariable=self.arrow10Var, width=3, validate='all', vcmd=vcmd)
+    arrow10Frame.pack()
+    arrow10Label.pack(anchor=W, side=LEFT, padx=(0,7))
+    arrow10Entry.pack(anchor=E)
+
+    bomb1Frame = Frame(itemList4)
+    bomb1Label = Label(bomb1Frame, text='Bomb (1)')
+    self.bomb1Var = StringVar(value='0')
+    bomb1Entry = Entry(bomb1Frame, textvariable=self.bomb1Var, width=3, validate='all', vcmd=vcmd)
+    bomb1Frame.pack()
+    bomb1Label.pack(anchor=W, side=LEFT, padx=(0,18))
+    bomb1Entry.pack(anchor=E)
+
+    bomb3Frame = Frame(itemList4)
+    bomb3Label = Label(bomb3Frame, text='Bombs (3)')
+    self.bomb3Var = StringVar(value='16')
+    bomb3Entry = Entry(bomb3Frame, textvariable=self.bomb3Var, width=3, validate='all', vcmd=vcmd)
+    bomb3Frame.pack()
+    bomb3Label.pack(anchor=W, side=LEFT, padx=(0,13))
+    bomb3Entry.pack(anchor=E)
+
+    bomb10Frame = Frame(itemList4)
+    bomb10Label = Label(bomb10Frame, text='Bombs (10)')
+    self.bomb10Var = StringVar(value='1')
+    bomb10Entry = Entry(bomb10Frame, textvariable=self.bomb10Var, width=3, validate='all', vcmd=vcmd)
+    bomb10Frame.pack()
+    bomb10Label.pack(anchor=W, side=LEFT, padx=(0,7))
+    bomb10Entry.pack(anchor=E)
+
+    rupee1Frame = Frame(itemList4)
+    rupee1Label = Label(rupee1Frame, text='Rupee (1)')
+    self.rupee1Var = StringVar(value='2')
+    rupee1Entry = Entry(rupee1Frame, textvariable=self.rupee1Var, width=3, validate='all', vcmd=vcmd)
+    rupee1Frame.pack()
+    rupee1Label.pack(anchor=W, side=LEFT, padx=(0,17))
+    rupee1Entry.pack(anchor=E)
+
+    rupee5Frame = Frame(itemList4)
+    rupee5Label = Label(rupee5Frame, text='Rupees (5)')
+    self.rupee5Var = StringVar(value='4')
+    rupee5Entry = Entry(rupee5Frame, textvariable=self.rupee5Var, width=3, validate='all', vcmd=vcmd)
+    rupee5Frame.pack()
+    rupee5Label.pack(anchor=W, side=LEFT, padx=(0,12))
+    rupee5Entry.pack(anchor=E)
+
+    rupee20Frame = Frame(itemList4)
+    rupee20Label = Label(rupee20Frame, text='Rupees (20)')
+    self.rupee20Var = StringVar(value='28')
+    rupee20Entry = Entry(rupee20Frame, textvariable=self.rupee20Var, width=3, validate='all', vcmd=vcmd)
+    rupee20Frame.pack()
+    rupee20Label.pack(anchor=W, side=LEFT, padx=(0,6))
+    rupee20Entry.pack(anchor=E)
+
+    rupee50Frame = Frame(itemList4)
+    rupee50Label = Label(rupee50Frame, text='Rupees (50)')
+    self.rupee50Var = StringVar(value='7')
+    rupee50Entry = Entry(rupee50Frame, textvariable=self.rupee50Var, width=3, validate='all', vcmd=vcmd)
+    rupee50Frame.pack()
+    rupee50Label.pack(anchor=W, side=LEFT, padx=(0,6))
+    rupee50Entry.pack(anchor=E)
+
+    rupee100Frame = Frame(itemList4)
+    rupee100Label = Label(rupee100Frame, text='Rupees (100)')
+    self.rupee100Var = StringVar(value='1')
+    rupee100Entry = Entry(rupee100Frame, textvariable=self.rupee100Var, width=3, validate='all', vcmd=vcmd)
+    rupee100Frame.pack()
+    rupee100Label.pack(anchor=W, side=LEFT, padx=(0,0))
+    rupee100Entry.pack(anchor=E)
+
+    rupee300Frame = Frame(itemList4)
+    rupee300Label = Label(rupee300Frame, text='Rupees (300)')
+    self.rupee300Var = StringVar(value='5')
+    rupee300Entry = Entry(rupee300Frame, textvariable=self.rupee300Var, width=3, validate='all', vcmd=vcmd)
+    rupee300Frame.pack()
+    rupee300Label.pack(anchor=W, side=LEFT, padx=(0,0))
+    rupee300Entry.pack(anchor=E)
+
+    blueclockFrame = Frame(itemList4)
+    blueclockLabel = Label(blueclockFrame, text='Blue Clock')
+    self.blueclockVar = StringVar(value='0')
+    blueclockEntry = Entry(blueclockFrame, textvariable=self.blueclockVar, width=3, validate='all', vcmd=vcmd)
+    blueclockFrame.pack()
+    blueclockLabel.pack(anchor=W, side=LEFT, padx=(0,11))
+    blueclockEntry.pack(anchor=E)
+
+    greenclockFrame = Frame(itemList4)
+    greenclockLabel = Label(greenclockFrame, text='Green Clock')
+    self.greenclockVar = StringVar(value='0')
+    greenclockEntry = Entry(greenclockFrame, textvariable=self.greenclockVar, width=3, validate='all', vcmd=vcmd)
+    greenclockFrame.pack()
+    greenclockLabel.pack(anchor=W, side=LEFT, padx=(0,3))
+    greenclockEntry.pack(anchor=E)
+
+    redclockFrame = Frame(itemList4)
+    redclockLabel = Label(redclockFrame, text='Red Clock')
+    self.redclockVar = StringVar(value='0')
+    redclockEntry = Entry(redclockFrame, textvariable=self.redclockVar, width=3, validate='all', vcmd=vcmd)
+    redclockFrame.pack()
+    redclockLabel.pack(anchor=W, side=LEFT, padx=(0,14))
+    redclockEntry.pack(anchor=E)
+
+    silverarrowFrame = Frame(itemList5)
+    silverarrowLabel = Label(silverarrowFrame, text='Silver Arrow')
+    self.silverarrowVar = StringVar(value='0')
+    silverarrowEntry = Entry(silverarrowFrame, textvariable=self.silverarrowVar, width=3, validate='all', vcmd=vcmd)
+    silverarrowFrame.pack()
+    silverarrowLabel.pack(anchor=W, side=LEFT, padx=(0,64))
+    silverarrowEntry.pack(anchor=E)
+
+    universalkeyFrame = Frame(itemList5)
+    universalkeyLabel = Label(universalkeyFrame, text='Universal Key')
+    self.universalkeyVar = StringVar(value='0')
+    universalkeyEntry = Entry(universalkeyFrame, textvariable=self.universalkeyVar, width=3, validate='all', vcmd=vcmd)
+    universalkeyFrame.pack()
+    universalkeyLabel.pack(anchor=W, side=LEFT, padx=(0,57))
+    universalkeyEntry.pack(anchor=E)
+
+    triforcepieceFrame = Frame(itemList5)
+    triforcepieceLabel = Label(triforcepieceFrame, text='Triforce Piece')
+    self.triforcepieceVar = StringVar(value='0')
+    triforcepieceEntry = Entry(triforcepieceFrame, textvariable=self.triforcepieceVar, width=3, validate='all', vcmd=vcmd)
+    triforcepieceFrame.pack()
+    triforcepieceLabel.pack(anchor=W, side=LEFT, padx=(0,55))
+    triforcepieceEntry.pack(anchor=E)
+
+    triforcecountFrame = Frame(itemList5)
+    triforcecountLabel = Label(triforcecountFrame, text='Triforce Pieces Required')
+    self.triforcecountVar = StringVar(value='0')
+    triforcecountEntry = Entry(triforcecountFrame, textvariable=self.triforcecountVar, width=3, validate='all', vcmd=vcmd)
+    triforcecountFrame.pack()
+    triforcecountLabel.pack(anchor=W, side=LEFT, padx=(0,0))
+    triforcecountEntry.pack(anchor=E)
+
+    triforceFrame = Frame(itemList5)
+    triforceLabel = Label(triforceFrame, text='Triforce (win game)')
+    self.triforceVar = StringVar(value='0')
+    triforceEntry = Entry(triforceFrame, textvariable=self.triforceVar, width=3, validate='all', vcmd=vcmd)
+    triforceFrame.pack()
+    triforceLabel.pack(anchor=W, side=LEFT, padx=(0,23))
+    triforceEntry.pack(anchor=E)
+
+    rupoorFrame = Frame(itemList5)
+    rupoorLabel = Label(rupoorFrame, text='Rupoor')
+    self.rupoorVar = StringVar(value='0')
+    rupoorEntry = Entry(rupoorFrame, textvariable=self.rupoorVar, width=3, validate='all', vcmd=vcmd)
+    rupoorFrame.pack()
+    rupoorLabel.pack(anchor=W, side=LEFT, padx=(0,87))
+    rupoorEntry.pack(anchor=E)
+
+    rupoorcostFrame = Frame(itemList5)
+    rupoorcostLabel = Label(rupoorcostFrame, text='Rupoor Cost')
+    self.rupoorcostVar = StringVar(value='10')
+    rupoorcostEntry = Entry(rupoorcostFrame, textvariable=self.rupoorcostVar, width=6, validate='all', vcmd=vcmd)
+    rupoorcostFrame.pack()
+    rupoorcostLabel.pack(anchor=W, side=LEFT, padx=(0,43))
+    rupoorcostEntry.pack(anchor=E)
+
+    itemList1.pack(side=LEFT, padx=(0,0))
+    itemList2.pack(side=LEFT, padx=(0,0))
+    itemList3.pack(side=LEFT, padx=(0,0))
+    itemList4.pack(side=LEFT, padx=(0,0))
+    itemList5.pack(side=LEFT, padx=(0,0))
+
+    return self
